move the storage response types to a module outsdie draft-storage
